Update UDP port scanning logic
Unless a UDP port reponds with data, we cannot verify
whether or not it is open. Use the following scenarios
to determine if a port is open, closed, or if its state
is unknown.

If you can write and read from a port it is open.
If you get an error writing to a port it is closed.
If you can write to a port but cannot read from it
it is unknown whether the port is open or closed.

Since UDP can fail, we retry each scan three times (this
could be increased).